requirements.txt
pika
pymongo
requests
beautifulsoup4
selenium
lxml
newspaper3k
language-tool-python
